Refactor fail_fast in preparation for interrupt support
